requirements.txt
akshare
flask
openai
pandas
pyarrow
pyyaml
pytest
